moved geolocation to api folder
